requirements.txt
typer>=0.12.0
rich>=13.7.0
scikit-learn>=1.4.0
numpy>=1.26.0
python-dateutil>=2.9.0
